+++ Server/CommandSystem/CommandCharacter/CommandCharacterGetStartMahjong.cpp
@@ -16,5 +16,5 @@ void CommandCharacterGetStartMahjong::execute()
 
 std::string CommandCharacterGetStartMahjong::showDebugInfo()
 {
-	COMMAND_DEBUG("mahjong : %s", MAHJONG_NAME[mMahjong].c_str());
+	COMMAND_DEBUG("mahjong : %s", MAHJONG_NAME[(int)mMahjong].c_str());
 }
