+++ Server/CommandSystem/CommandCharacter/CommandCharacterNotifyOtherPlayerDrop.cpp
@@ -21,5 +21,5 @@ void CommandCharacterNotifyOtherPlayerDrop::execute()
 
 std::string CommandCharacterNotifyOtherPlayerDrop::showDebugInfo()
 {
-	COMMAND_DEBUG("player guid : %d, mahjong : %s", mPlayerGUID, MAHJONG_NAME[mMahjong].c_str());
+	COMMAND_DEBUG("player guid : %d, mahjong : %s", (int)mPlayerGUID, MAHJONG_NAME[(int)mMahjong].c_str());
 }
