+++ Server/CommandSystem/CommandCharacter/CommandCharacterNotifyOtherPlayerGetStartMahjong.cpp
@@ -15,5 +15,5 @@ void CommandCharacterNotifyOtherPlayerGetStartMahjong::execute()
 
 std::string CommandCharacterNotifyOtherPlayerGetStartMahjong::showDebugInfo()
 {
-	COMMAND_DEBUG("player guid : %d, mahjong : %s", mPlayerGUID, MAHJONG_NAME[mMahjong].c_str());
+	COMMAND_DEBUG("player guid : %d, mahjong : %s", (int)mPlayerGUID, MAHJONG_NAME[(int)mMahjong].c_str());
 }
