+++ Server/CommandSystem/CommandCharacter/CommandCharacterShowHua.cpp
@@ -17,5 +17,5 @@ void CommandCharacterShowHua::execute()
 
 std::string CommandCharacterShowHua::showDebugInfo()
 {
-	COMMAND_DEBUG("index : %d, mahjong : %s", mIndex, MAHJONG_NAME[mMahjong].c_str());
+	COMMAND_DEBUG("index : %d, mahjong : %s", mIndex, MAHJONG_NAME[(int)mMahjong].c_str());
 }
